extracted implementation of AssetNode methods into a base abstract class
